+++ app/src/main/java/com/example/onepiece_android/ServerApi.java
@@ -48,7 +48,6 @@ public interface ServerApi {
             @Query("place") String place
     );
 
-
     @POST("/tag/{feedId}")
     Call<TagResponse> createTag(
             @Path("boardId") long feedId,
