+++ app/src/main/res/layout/activity_post_modify.xml
@@ -100,22 +100,21 @@
         app:layout_constraintTop_toBottomOf="@+id/tv_checktag" />
 
 
-    <TextView
-        android:id="@+id/tv_plus"
+    <ImageView
+        android:id="@+id/img_delete"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_marginEnd="20dp"
-        android:fontFamily="@font/orbit"
-        android:text="+"
-        android:textColor="@color/gray_700"
-        android:textSize="32sp"
+        android:src="@drawable/delete"
         app:layout_constraintBottom_toBottomOf="@+id/et_tag"
         app:layout_constraintEnd_toEndOf="@+id/et_tag"
-        app:layout_constraintTop_toTopOf="@+id/et_tag" />
+        app:layout_constraintTop_toTopOf="@+id/et_tag"
+        android:clickable="true"
+        android:focusable="true"/>
 
     <LinearLayout
         android:id="@+id/linearLayout"
-        android:layout_width="match_parent"
+        android:layout_width="343dp"
         android:layout_height="wrap_content"
         android:layout_marginStart="37dp"
         android:layout_marginTop="8dp"
